% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{add.fish.dates}
\alias{add.fish.dates}
\title{Add better formatted dates}
\usage{
add.fish.dates(.data)
}
\description{
This function takes dates as formatted by the NZFFD and converts them into actual date columns
}
