% Generated by roxygen2: do not edit by hand
% Please edit documentation in functions/functions.r
\name{add.fish.ibi}
\alias{add.fish.ibi}
\title{Add combined IBI score}
\usage{
add.fish.ibi(.data)
}
\description{
Takes the individual metric scores and combines them to form the final IBI score
(continuous)
}
