% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bad.R
\name{bad}
\alias{bad}
\title{Returns parameters from a Stan object which are not converged}
\usage{
bad(fit, r.hat.level = 1.2, plot = FALSE)
}
\arguments{
\item{fit}{Model object from Stan}

\item{r.hat.level}{Argument controling the level for Rhat. Defaults to 1.2}

\item{plot}{Display trace plots of these parameters. (not working)}
}
\description{
Returns the parameters from a Stan object which are not
  converged, based on Rhat.
}
\examples{
\dontrun{
library(rstan)
scode <- "
parameters {
real y[2];
}
model {
y[1] ~ normal(0, 1);
y[2] ~ double_exponential(0, 2);
}
"
fit <- stan(model_code = scode, iter = 10, verbose = FALSE)
bad(fit)
}
}
\author{
Michael J. Dodrill, \email{mdodrill@usgs.gov}
}
