% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_summary.R
\name{bayes_summary}
\alias{bayes_summary}
\title{Calculate mean, 95% Lower and Upper CRI, from JAGS and Stan}
\usage{
bayes_summary(fit, par.name)
}
\arguments{
\item{fit}{Model object from JAGS or Stan}

\item{par.name}{Parameter of intrest to return summary stats}
}
\description{
Takes an fitted Bayesian model object from either JAGS or Stan and returns summary stats.
}
\examples{
Add these in...
}
\author{
Michael J. Dodrill, \email{mdodrill@usgs.gov}
}
