% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{cut.fish.ibi}
\alias{cut.fish.ibi}
\title{Cut IBI score}
\usage{
\method{cut}{fish.ibi}(.data)
}
\description{
Takes the continuous IBI overall score and cuts it into the three categories
originally proposed by Joy and Death (2004)
}
