% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fao.R
\name{drop_non_english_cols}
\alias{drop_non_english_cols}
\title{Drop non-English language columns from FAO lookup tables}
\usage{
drop_non_english_cols(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame without Spanish, Russian, Arabic, Chinese,
and French columns.
}
\description{
Drop non-English language columns from FAO lookup tables
}
\keyword{internal}
