% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{effort}
\alias{effort}
\title{Correct effort for hoop data}
\usage{
effort(x)
}
\arguments{
\item{x}{is the data.frame, ex. sample table with the correct columns names}
}
\description{
This function can be applied to hoop net data to produce the correct effort.
}
\details{
Function needs columns: start_datetime, end_datetime, effort, total_hours and trip_id
}
\examples{
Need to add...
}
\author{
Michael J. Dodrill, \email{mdodrill@usgs.gov}
}
