% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fao.R
\name{find_fao_capture_file}
\alias{find_fao_capture_file}
\title{Find a file in the extracted FAO capture dataset}
\usage{
find_fao_capture_file(path, filename)
}
\arguments{
\item{path}{Path to extracted FAO capture directory.}

\item{filename}{File name to locate.}
}
\value{
Full path to file.
}
\description{
Find a file in the extracted FAO capture dataset
}
\keyword{internal}
