% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprint.R
\name{fish_rois}
\alias{fish_rois}
\title{Fetch the ROI hierarchy for fish2}
\usage{
fish_rois(
  root = NULL,
  rval = c("edgelist", "graph"),
  cache = FALSE,
  conn = NULL,
  ...,
  dataset = fish_default_dataset()
)
}
\arguments{
\item{root}{Character vector specifying a root ROI. The default
(\code{NULL}) returns the whole hierarchy.}

\item{rval}{Whether to return an edge list \code{data.frame} (default) or an
\code{igraph} object.}

\item{cache}{if \code{TRUE} will use memoisation to cache the result of the
call for 24 hours.}

\item{conn}{Optional, a \code{neuprint_connection} object. Defaults to
\code{\link{fish_neuprint}} to ensure that the query targets fish2.}

\item{...}{Additional arguments passed to
\code{\link[neuprintr]{neuprint_ROI_hierarchy}}.}

\item{dataset}{optional, a dataset you want to query. If \code{NULL}, the
default specified by your R environ file is used or, failing that the
current connection, is used. See \code{\link[neuprintr]{neuprint_login}} for details.}
}
\value{
The ROI hierarchy as either an edge list or \code{igraph} object.
}
\description{
Fetch the ROI hierarchy for fish2
}
\examples{
\donttest{
head(fish_rois())
}
}
\seealso{
\code{\link[neuprintr]{neuprint_ROI_hierarchy}},
  \code{\link{fish_roi_meshes}}

Other 3d-meshes-skeletons:
\code{\link[=fish_roi_meshes]{fish_roi_meshes()}},
\code{\link[=read_fish_meshes]{read_fish_meshes()}},
\code{\link[=read_fish_neurons]{read_fish_neurons()}}
}
\concept{3d-meshes-skeletons}
