% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishr-package.R
\docType{package}
\name{fishr-package}
\alias{fishr}
\alias{fishr-package}
\title{fishr: Calculate standard metrics for fisheries (One Line, Title Case)}
\description{
Bla bla bla bla.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/davidsean/fishr}
  \item Report bugs at \url{https://github.com/davidsean/fishr/issues}
}

}
\author{
\strong{Maintainer}: David Sean-Fortin \email{david.sean-fortin@dfo-mpo.gc.ca}

}
\keyword{internal}
