% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{horsefly_sample}
\alias{horsefly_sample}
\title{Rainbow Trout gill net survey data from Horsefly Lake}
\format{
A data frame with columns:
\itemize{
\item \code{set_number} Integer. Net set identifier.
\item \code{catch} Integer. Number of fish caught per set.
\item \code{effort} Integer. Number of net-nights per set.
}
}
\source{
Fictional data for illustration purposes.
}
\usage{
horsefly_sample
}
\description{
Catch and effort aggregated to the net set level from a fictional
gill net survey of Rainbow Trout (\emph{Oncorhynchus mykiss}) at
Horsefly Lake, British Columbia.
}
\examples{
cpue(horsefly_sample)
}
\keyword{datasets}
