% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{organize}
\alias{organize}
\title{Organizes fitted model objects from JAGS and Stan}
\usage{
organize(fit, par.name, mcmc.out = FALSE)
}
\arguments{
\item{fit}{Model object from JAGS or Stan}

\item{par.name}{Parameter of intrest to organize into a dataframe}

\item{mcmc.out}{Logical, whether the full dataframe with all values should be
returned. Defaults to FALSE.}
}
\description{
Takes an fitted Bayesian model object from eith JAGS or Stan and returns a dataframe.
}
\examples{
Add these in...
}
\author{
Michael J. Dodrill, \email{mdodrill@usgs.gov}
}
