% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fao.R
\name{plot_top_species_country}
\alias{plot_top_species_country}
\title{Plot top captured species for a selected country}
\usage{
plot_top_species_country(data, country, year, n = 10, fill = "#0a4875")
}
\arguments{
\item{data}{A data frame returned by \code{load_fao_capture()}.}

\item{country}{Country name in English, matching \code{Country_Name_En}.}

\item{year}{Year to plot.}

\item{n}{Number of top species to show.}

\item{fill}{Fill colour for bars.}
}
\value{
A ggplot object.
}
\description{
Plot top captured species for a selected country
}
