% Generated by roxygen2: do not edit by hand
% Please edit documentation in functions/functions.r
\name{qr.construct}
\alias{qr.construct}
\title{Fit quantile regression}
\usage{
\method{qr}{construct}(y, x, data = site_metrics_all)
}
\description{
This function fits a quantile regression for scoring of metrics. This improves
upon methods which require fitting by eye.
}
