% Generated by roxygen2: do not edit by hand
% Please edit documentation in functions/functions.r
\name{qr.score}
\alias{qr.score}
\title{Score metrics according to quantile regression}
\usage{
\method{qr}{score}(x, y, qr)
}
\description{
Once you've fit a quantile regression, you then need to score it
}
