% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{read_fish_meshes}
\alias{read_fish_meshes}
\title{Read meshes for fish2 body ids}
\usage{
read_fish_meshes(ids, node = "neutu", units = c("nm", "raw", "microns"), ...)
}
\arguments{
\item{ids}{One or more body ids (numeric or character), or a query string
accepted by \code{\link{fish_ids}}.}

\item{node}{The DVID node (UUID) to query. The default value of 'neutu' uses
the active neutu node i.e. normally the most up to date.}

\item{units}{One of \code{"nm"} (default), \code{"raw"} (nominal voxel size),
or \code{"microns"}.}

\item{...}{Additional arguments passed to \code{httr::GET}.}
}
\value{
A \code{\link[nat:neuronlist]{neuronlist}} containing one or more
  \code{mesh3d} objects.
}
\description{
Fetches pre-computed neuroglancer meshes from the fish2 DVID
  server.
}
\details{
It seems that the nominal voxel dimensions of the fish2 dataset are
  16 x 16 x 15 nm. Neuron meshes are already returned from DVID in physical
  nm scale but this voxel dimension is required to turn raw neuroglancer
  coordinates into nm. This means that if you click on a neuron in
  neuroglancer, the recorded location will match if you have chosen
  `units="nm"`. Alternatively you must multiply point locations from
  neuroglancer by c(16,16,15) to obtain points in nm space.
}
\examples{
\dontrun{
# pick 10 RGC meshes to read in
rgcs <- read_fish_meshes(fish_ids('RGC')[1:10])
plot3d(rgcs)
}
}
\seealso{
Other 3d-meshes-skeletons:
\code{\link[=fish_roi_meshes]{fish_roi_meshes()}},
\code{\link[=fish_rois]{fish_rois()}},
\code{\link[=read_fish_neurons]{read_fish_neurons()}}
}
\concept{3d-meshes-skeletons}
