% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coconat.R
\name{register_fish_coconat}
\alias{register_fish_coconat}
\title{Register fish2 dataset for coconatfly}
\usage{
register_fish_coconat(showerror = TRUE)
}
\arguments{
\item{showerror}{Logical; when \code{FALSE}, return invisibly if
dependencies are missing.}
}
\value{
Invisible \code{NULL}.
}
\description{
Register fish2 dataset adapters for use with
  \href{https://natverse.org/coconatfly}{coconatfly}.
}
\examples{
\dontrun{
register_fish_coconat()
coconatfly::cf_meta(coconatfly::cf_ids(fish2 = "RGC"))
}
}
\seealso{
Other data-queries:
\code{\link[=fish_connection_table]{fish_connection_table()}},
\code{\link[=fish_ids]{fish_ids()}},
\code{\link[=fish_neuprint_meta]{fish_neuprint_meta()}},
\code{\link[=fish_somapos]{fish_somapos()}}
}
\concept{data-queries}
