% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_trace.R
\name{stan_trace}
\alias{stan_trace}
\title{A basic function}
\usage{
stan_trace(fit, par.name, number, same.scale = FALSE)
}
\arguments{
\item{test}{Tests if the function is working. Defaults to TRUE.}
}
\description{
This is a test function I wrote.
}
\examples{
test()
}
\author{
Michael J. Dodrill, \email{mdodrill@usgs.gov}
}
