% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_db.R
\name{update_db}
\alias{update_db}
\title{A basic function}
\usage{
update_db()
}
\arguments{
\item{test}{Tests if the function is working. Defaults to TRUE.}
}
\description{
This is a test function I wrote.
}
\examples{
test()
}
\author{
Michael J. Dodrill, \email{mdodrill@usgs.gov}
}
