% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bad.R
\name{bad}
\alias{bad}
\title{Returns parameters from a Stan object which are not converged}
\usage{
bad(fit, r.hat.level = 1.2, plot = FALSE)
}
\arguments{
\item{fit}{Model object from Stan}

\item{r.hat.level}{Argument controling the level for Rhat. Defaults to 1.2}

\item{plot}{Display trace plots of these parameters. (not working)}
}
\description{
Returns the parameters from a Stan object which are not
  converged, based on Rhat.
}
\examples{
Need to add these
}
\author{
Michael J. Dodrill, \email{mdodrill@usgs.gov}
}
