% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{biomass_index}
\alias{biomass_index}
\title{Calculate biomass index}
\usage{
biomass_index(cpue = NULL, area_swept, catch = NULL, effort = NULL, ...)
}
\arguments{
\item{cpue}{Numeric vector of CPUE values}

\item{area_swept}{Numeric vector of area swept (e.g., km^2)}

\item{catch}{Numeric}

\item{effort}{Numeric}

\item{...}{arguments passed to CPUE}
}
\value{
A numeric vector of the biomass index
}
\description{
Calculation of biomass index from the CPUE and area swept
}
\examples{
salmon_cpue <- cpue(10, 2)
area_swept <- 10
biomass_index(salmon_cpue, area_swept)
}
