% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{biomass_index}
\alias{biomass_index}
\title{Calculate Biomass Index}
\usage{
biomass_index(
  cpue = NULL,
  area_swept,
  catch = NULL,
  effort = NULL,
  verbose = getOption("fishr.verbose", default = FALSE),
  ...
)
}
\arguments{
\item{cpue}{Numeric vector of CPUE values}

\item{area_swept}{Numeric vector of area swept (e.g., km²)}

\item{catch}{Numeric value of catch}

\item{effort}{Numeric value of effort}

\item{verbose}{Logical indicating whether to print processing messages (default is FALSE, also accept the value of 'fishr.verbose')}

\item{...}{Arguments passed to cpue()}
}
\value{
A numeric vector of biomass index values
}
\description{
Calculates biomass index from CPUE and area swept.
}
\examples{
salmon_cpue <- cpue(catch = 2, effort = 2)
biomass_index(cpue = salmon_cpue, area_swept = 5)
}
