% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{biomass_index}
\alias{biomass_index}
\title{Calculate Biomass Index
Calculate biomass index from CPUE and area swept}
\usage{
biomass_index(cpue, area_swept)
}
\arguments{
\item{cpue}{Numeric vector of CPUE values}

\item{area_swept}{Numeric vector of area swept (eg km squared)}
}
\value{
Numeric vector of biomass index values
}
\description{
Calculate Biomass Index
Calculate biomass index from CPUE and area swept
}
\examples{
salmon_cpue <- cpue(catch = 2, effort = 2)
biomass_index(cpue = salmon_cpue, area_swept = 5)
}
