% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass-index.R
\name{biomass_index}
\alias{biomass_index}
\title{Calculate biomass index}
\usage{
biomass_index(cpue = NULL, area_swept, catch = NULL, effort = NULL, ...)
}
\arguments{
\item{cpue}{Numeric vector of CPUE values}

\item{area_swept}{Numeric vector of area swept ( e.g., km^2)}

\item{catch}{Numeric}

\item{effort}{Numeric}

\item{...}{arguments passed to CPUE}
}
\value{
A numeric vector of biomass index values
}
\description{
Calculate biomass index
}
\examples{
salmon_cpue <- cpue(catch = 2, effort =2 )
biomass_index( salmon_cpue, 3)
}
