% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{biomass_index}
\alias{biomass_index}
\title{Biomass Index}
\usage{
biomass_index(cpue, area_swept)
}
\arguments{
\item{cpue}{numeric catch per unit effort}

\item{area_swept}{numeric area}
}
\value{
numeric biomass
}
\description{
Biomass Index
}
\examples{
biomass_index(10, 2)
}
