% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomasse.R
\name{biomass_index}
\alias{biomass_index}
\title{Calculer biomasse index}
\usage{
biomass_index(cpue, area_sewpt)
}
\arguments{
\item{cpue}{nomerci vector}

\item{area_sewpt}{numeric vector de l'aire balayée}
}
\value{
value
}
\description{
utilise l'aire balbayées
}
\examples{
biomass_index(10, 10)
}
