% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{biomass_index}
\alias{biomass_index}
\title{Calculate Biomass Index}
\usage{
biomass_index(cpue, area_swept)
}
\arguments{
\item{cpue}{Numberic vector of CPUE values}

\item{area_swept}{Numeric vector of area swept (eg km^2)}
}
\value{
returns numeric vector of biomass index values
}
\description{
Calculates biomass index from CPUE and area swept
}
\examples{
salmon_cpue <- cpue(catch = 2, effort = 2)
biomass_index(salmon_cpue, 5)
}
