% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{biomass_index}
\alias{biomass_index}
\title{Calculate biomass index}
\usage{
biomass_index(cpue, area_swept)
}
\arguments{
\item{cpue}{Numeric vector of CPUE values}

\item{area_swept}{Numeric vector of area swept (e.g., km^2)}
}
\value{
A  numeric vector of biomass index values
}
\description{
Calculate biomass index from CPUE and area swept
}
\examples{
cod_cpue <- cpue(catch=100, effort=20)
biomass_index(cod_cpue, 5)
}
