% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{biomass_index}
\alias{biomass_index}
\title{Calculate biomass index}
\usage{
biomass_index(
  cpue = NULL,
  area_swept,
  catch = NULL,
  effort = NULL,
  verbose = getOption("fishr.verbose", default = FALSE),
  ...
)
}
\arguments{
\item{cpue}{Numeric vector of CPUE values (units: kg/hr).}

\item{area_swept}{Numeric vector of area swept in \strong{km^2^}.}

\item{catch}{Numeric vector of catch values}

\item{effort}{Numeric vector of effort values}

\item{verbose}{Logical, if TRUE print processing info (default is FALSE, also accepts the value of )}

\item{...}{
  Arguments passed on to \code{\link[=cpue]{cpue}}
  \describe{
    \item{\code{gear_factor}}{Numeric adjustment for fishing gear corrections}
    \item{\code{method}}{Character; one of "ratio" or "log"}
  }}
}
\value{
A  numeric vector of biomass index values
}
\description{
Calculate biomass index from CPUE and area swept
}
\examples{
cod_cpue <- cpue(catch=100, effort=20)
biomass_index(cod_cpue, 5)
}
