% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{biomass_index}
\alias{biomass_index}
\title{Calculate biomass index}
\usage{
biomass_index(
  cpue = NULL,
  area_swept,
  catch = NULL,
  effort = NULL,
  verbose = getOption("fishr.verbose", default = FALSE),
  ...
)
}
\arguments{
\item{cpue}{Numeric vector of CPUE values}

\item{area_swept}{Numeric vector of area swept (e.g., km^2)}

\item{catch}{catch value numeric}

\item{effort}{effort value numeric}

\item{verbose}{Logical, if TRUE print processing info (default is FALSE, also accepts the value of )}

\item{...}{arguments passed to cpue function}
}
\value{
A  numeric vector of biomass index values
}
\description{
Calculate biomass index from CPUE and area swept
}
\examples{
cod_cpue <- cpue(catch=100, effort=20)
biomass_index(cod_cpue, 5)
}
