% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{biomass_index}
\alias{biomass_index}
\title{Calculate biomass index}
\usage{
biomass_index(
  cpue = NULL,
  area_swept,
  catch = NULL,
  effort = NULL,
  verbose = getOption("fishr.verbose", default = FALSE),
  ...
)
}
\arguments{
\item{cpue}{Numeric vector of CPUE values}

\item{area_swept}{Numeric vector of area swept (e.g., \strong{km^2^})}

\item{catch}{Numeric vector of catch (e.g. \strong{kg}) or a data.frame with "catch" and "effort" columns}

\item{effort}{Numeric vector of effort (e.g. \strong{hours})}

\item{verbose}{Logical, indicating whether to print processing messages (default is FALSE, also accepts value of \code{fishr.verbose})}

\item{...}{
  Arguments passed on to \code{\link[=cpue]{cpue}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
A numeric vector of biomass index values
}
\description{
Calculates biomass from cpue and area fished.
}
\details{
Two modes of use
\itemize{
\item Provide \code{cpue} directly
\item Provide \code{catch} and \code{effort} separately
}
}
\examples{
salmon_cpue <- cpue(catch = 2, effort = 2)
biomass_index(salmon_cpue, 5)
}
\seealso{
See \code{\link[=cpue]{cpue()}} for more details.
}
