% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{biomass_index}
\alias{biomass_index}
\title{Calculate biomass index}
\usage{
biomass_index(
  cpue = NULL,
  area_swept,
  catch = NULL,
  effort = NULL,
  verbose = getOption("fishr.verbose", default = FALSE),
  ...
)
}
\arguments{
\item{cpue}{Numeric vector of CPUE values}

\item{area_swept}{Numeric vector of area swept (e.g., km^2)}

\item{catch}{Numeric vector of catch}

\item{effort}{Numeric vector of effort}

\item{verbose}{Logical, indicating whether to print processing messages (default is FALSE, also accepts value of \code{fishr.verbose})}

\item{...}{arguments passed to cpue}
}
\value{
A numeric vector of biomass index values
}
\description{
Calculates biomass from cpue and area fished
}
\examples{
salmon_cpue <- cpue(catch = 2, effort = 2)
biomass_index(salmon_cpue, 5)
}
