% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_fish_db.R
\name{connect_fish_db}
\alias{connect_fish_db}
\title{A basic function}
\usage{
connect_fish_db(update.db = FALSE)
}
\arguments{
\item{test}{Tests if the function is working. Defaults to TRUE.}
}
\description{
This is a test function I wrote.
}
\examples{
test()
}
\author{
Michael J. Dodrill, \email{mdodrill@usgs.gov}
}
