% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{cpue}
\alias{cpue}
\title{Calculate Catch per unit effort (CPUE)}
\usage{
cpue(
  catch,
  effort,
  gear_factor = 1,
  method = c("ratio", "log"),
  verbose = getOption("fishr.verbose", default = FALSE)
)
}
\arguments{
\item{catch}{Numeric vector of catch}

\item{effort}{Numeric vector of effort}

\item{gear_factor}{Numeric adjustment for gear standardize}

\item{method}{}

\item{verbose}{Logical value to show messages (default is FALSE, use fishr.verbose option to set globally)}
}
\value{
a numeric, vector of CPUE values
}
\description{
Calculate Catch per unit effort (CPUE) as catch divided by effort, optionally adjusting for gear standardization.
}
\examples{
cpue(100,10)
cpue(100, 10, gear_factor=0.5)
}
