% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupe.R
\name{cpue}
\alias{cpue}
\alias{cpue.numeric}
\alias{cpue.data.frame}
\alias{cpue.default}
\title{Calculate Catch Per Unit Effort (CPUE)}
\usage{
cpue(catch, ...)

\method{cpue}{numeric}(
  catch,
  effort,
  gear_factor = 1,
  method = c("ratio", "log"),
  verbose = getOption("fishr.verbose", default = FALSE),
  ...
)

\method{cpue}{data.frame}(
  catch,
  gear_factor = 1,
  method = c("ratio", "log"),
  verbose = getOption("fishr.verbose", FALSE),
  ...
)

\method{cpue}{default}(catch, ...)
}
\arguments{
\item{catch}{Numeric vector of catch (e.g. kg) or a data frame of "catch" and "effort" columns}

\item{...}{Additional arguments passed on to methods}

\item{effort}{Numeric vector of effort (e.g. hours)}

\item{gear_factor}{Numeric adjustment for gear standard (defaults is 1)}

\item{method}{Character: one of "ratio" or "log"; will take "ratio" as default}

\item{verbose}{Logical indicating whether to print processing messages (default is FALSE, also accepts the value of 'fishr.verbose')}
}
\value{
Numeric vector of CPUE values of the class 'my_cpue'
}
\description{
Calculates CPUE from catch and effort data, with optional gear standardization.
}
\examples{
cpue(100, 10)
cpue(100, 10, gear_factor = 0.5)
}
