% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupe.R
\name{cpue}
\alias{cpue}
\alias{cpue.numeric}
\alias{cpue.data.frame}
\alias{cpue.default}
\title{Calculate Catch Per Unit Effort (CPUE)}
\usage{
cpue(catch, ...)

\method{cpue}{numeric}(
  catch,
  effort,
  gear_factor = NULL,
  gear_type = "nordic_gillnet",
  method = c("ratio", "log"),
  verbose = getOption("fishr.verbose", default = FALSE)
)

\method{cpue}{data.frame}(
  catch,
  gear_factor = 1,
  method = c("ratio", "log"),
  verbose = getOption("fishr.verbose", FALSE),
  ...
)

\method{cpue}{default}(catch, ...)
}
\arguments{
\item{catch}{Input data: a numeric vector of catch values, or a data frame
containing catch and effort columns.}

\item{...}{Additional arguments passed to methods.}

\item{effort}{Numeric vector of effort (e.g. hours)}

\item{gear_factor}{Numeric adjustment for gear standard (defaults is 1)}

\item{method}{Character: one of "ratio" or "log"; will take "ratio" as default}

\item{verbose}{Logical indicating whether to print processing messages (default is FALSE, also accepts the value of 'fishr.verbose')}
}
\value{
Numeric vector of CPUE values
}
\description{
Calculates CPUE from catch and effort data, with optional gear
standardization. Supports ratio and log-transformed methods.
}
\examples{
cpue(100, 10)
cpue(100, 10, gear_factor = 0.5)
}
