% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{cpue}
\alias{cpue}
\title{Calculate Catch per unit Effort (CPUE)}
\usage{
cpue(catch, effort, gear_factor = 1)
}
\arguments{
\item{catch}{Numeric vector of catch (e.g., kg)}

\item{effort}{Numeric vector of effort (e.g., hours)}

\item{gear_factor}{Numeric adjusment fo gear standardization (default is 1)}
}
\value{
A numeric vector of CPUE values
}
\description{
Calculate Catch per unit Effort (CPUE)
}
\examples{
cpue(100,10)
cpue(100,10,gear_factor=0.5)

}
