% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{cpue}
\alias{cpue}
\title{Calculate catch per unit effort (CPUE)}
\usage{
cpue(catch, effort, gear_factor = 1)
}
\arguments{
\item{catch}{Numeric vector of catch}

\item{effort}{Numeric vector of effort}

\item{gear_factor}{Numeric vector for gear standardize}
}
\value{
Numeric vector of CPUE values
}
\description{
Calculate catch per unit effort (CPUE)
}
\examples{
cpue(100,10)
cpue(100,10, gear_factor = 1.5)
}
