% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{cpue}
\alias{cpue}
\title{Calculate catch per unit effort (CPUE)}
\usage{
cpue(
  catch,
  effort,
  gear_factor = 1,
  method = c("ratio", "log"),
  verbose = getOption("fishr.verbose", default = FALSE)
)
}
\arguments{
\item{catch}{Numeric vector of catch}

\item{effort}{Numeric vector of effort}

\item{gear_factor}{Numeric vector for gear standardize}

\item{method}{Character; one of "ratio" or "log".}

\item{verbose}{Logical, if TRUE, prints messaging. Default is FALSE}
}
\value{
Numeric vector of CPUE values
}
\description{
Calculate catch per unit effort (CPUE)
}
\examples{
cpue(100,10)
cpue(100,10, gear_factor = 1.5)
}
