% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{cpue}
\alias{cpue}
\alias{cpue.numeric}
\alias{cpue.data.frame}
\alias{cpue.default}
\title{Calculate catch per unit effort (CPUE)}
\usage{
cpue(catch, ...)

\method{cpue}{numeric}(
  catch,
  effort,
  gear_factor = 1,
  method = c("ratio", "log"),
  verbose = getOption("fishr.verbose", default = FALSE),
  ...
)

\method{cpue}{data.frame}(
  catch,
  gear_factor = 1,
  method = c("ratio", "log"),
  verbose = getOption("fishr.verbose", FALSE),
  ...
)

\method{cpue}{default}(catch, ...)
}
\arguments{
\item{catch}{Numeric vector of catch, or data.frame with "catch" and "effort" columns}

\item{...}{}

\item{effort}{Numeric vector of effort}

\item{gear_factor}{Numeric vector for gear standardize}

\item{method}{Character; one of "ratio" or "log".}

\item{verbose}{Logical, if TRUE, prints messaging. Default is FALSE}
}
\value{
Numeric vector of CPUE values
}
\description{
Calculate catch per unit effort (CPUE)
}
\examples{
cpue(100,10)
cpue(100,10, gear_factor = 1.5)
}
