% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{cpue}
\alias{cpue}
\title{Calculate something}
\usage{
cpue(
  catch,
  effort,
  gear_factor = 1,
  method = c("ratio", "log"),
  verbose = getOption("fishr.verbose", default = FALSE)
)
}
\arguments{
\item{catch}{Numerci vector of catch}

\item{effort}{numeric}

\item{gear_factor}{Numeric}

\item{method}{Character; one of "ration" or "log"}

\item{verbose}{logical}
}
\value{
someting vector of numeric
}
\description{
Calculate something
}
\examples{
cpue(100,10)
cpue(100, 10, gear_factor=0.5)

}
