% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{cpue}
\alias{cpue}
\title{Calculate Catch Per unit effort}
\usage{
cpue(
  catch,
  effort,
  gear_factor = 1,
  method = c("ratio", "log"),
  verbose = getOption("fishr.verbose", FALSE)
)
}
\arguments{
\item{catch}{numeric vector of catch}

\item{effort}{numeric vector of effort}

\item{gear_factor}{numeric adjustment for gear standard}

\item{method}{}
}
\value{
numeric vector of CPUE values
}
\description{
Calculate Catch Per unit effort
}
\examples{
cpue(100,10)
cpue(100, 10, gear_factor = .5)
}
