% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{cpue}
\alias{cpue}
\title{Calculate catch per unit of effort (CPUE)}
\usage{
cpue(
  catch,
  effort,
  gear_factor = 1,
  method = c("ratio", "log"),
  verbose = getOption("fishr.verbose", default = FALSE)
)
}
\arguments{
\item{catch}{Numeric vector of catch values}

\item{effort}{Numeric vector of effort values}

\item{gear_factor}{Numeric adjustment for fishing gear corrections}

\item{method}{Character; one of "ratio" or "log"}

\item{verbose}{Logical, if TRUE print processing info (default is FALSE, also accepts the value of )}
}
\value{
A numeric vector of CPUE values
}
\description{
Calculate catch per unit of effort (CPUE)
}
\examples{
cpue(100, 10)
cpue(100, 10, gear_factor=2/3)
}
