% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{cpue}
\alias{cpue}
\title{Calculate Catch Per Unit Effort (CPUE)}
\usage{
cpue(
  catch,
  effort,
  gear_factor = 1,
  method = c("ratio", "log"),
  verbose = getOption("fishr.verbose", default = FALSE)
)
}
\arguments{
\item{catch}{Numeric vector of catch (e.g., kg)}

\item{effort}{Numeric vector of effort (e.g., hours)}

\item{gear_factor}{Numeric adjustment for gear standardization (default is 1)}

\item{method}{Character; one of "ratio" or "log"}

\item{verbose}{Logical indicating whether to print processing messages (default is FALSE, also accept value of "fish.verbose")}
}
\value{
A numeric vector of CPUE values
}
\description{
Calculates CPUE from catch and effort data, with optional gear standardization.
}
\examples{
cpue(100, 10)
cpue(100, 10, gear_factor = 0.5)
}
