% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{cpue}
\alias{cpue}
\alias{cpue.numeric}
\alias{cpue.data.frame}
\alias{cpue.default}
\title{Calculate catch per unit effort (CPUE)}
\usage{
cpue(catch, ...)

\method{cpue}{numeric}(
  catch,
  effort,
  gear_factor = 1,
  method = c("ratio", "log"),
  verbose = getOption("fishr.verbose", default = FALSE),
  ...
)

\method{cpue}{data.frame}(
  catch,
  gear_factor = 1,
  method = c("ratio", "log"),
  verbose = getOption("fishr.verbose", default = FALSE),
  ...
)

\method{cpue}{default}(catch, ...)
}
\arguments{
\item{catch}{Numeric vector of catch (e.g. \strong{kg}) or a data.frame with "catch" and "effort" columns}

\item{...}{Additional argument passed on to methods}

\item{effort}{Numeric vector of effort (e.g. \strong{hours})}

\item{gear_factor}{Numeric adjustment for gear standardize}

\item{method}{Character; one of \code{"ratio"} or \code{"log"}}

\item{verbose}{Logical, indicating whether to print processing messages (default is FALSE, also accepts value of \code{fishr.verbose})}
}
\value{
A numeric vector of CPUE
}
\description{
Calculates CPUE from catch and effort data, with optional gear standardization.
}
\examples{
cpue(100,10)
cpue(100,10, gear_factor = 0.5)
}
