% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprint.R
\name{fish_connection_table}
\alias{fish_connection_table}
\title{Connectivity query for fish2 neurons}
\usage{
fish_connection_table(
  ids,
  partners = c("inputs", "outputs"),
  moredetails = FALSE,
  summary = FALSE,
  threshold = 1L,
  roi = NULL,
  by.roi = FALSE,
  conn = NULL,
  ...,
  dataset = fish_default_dataset()
)
}
\arguments{
\item{ids}{A set of body ids (see \code{\link{fish_ids}} for a range of ways
to specify these).}

\item{partners}{Either inputs or outputs. Redundant with \code{prepost}, but
probably clearer.}

\item{moredetails}{\code{FALSE} (default) to use only the columns returned by
\code{neuprint_connection_table}, \code{TRUE} to join all additional fields
from \code{\link{fish_neuprint_meta}}, or a character vector naming
specific extra fields (e.g. \code{c("group")}).}

\item{summary}{When \code{TRUE} and more than one query neuron is given,
summarises connectivity grouped by partner.}

\item{threshold}{Only return partners >= to an integer value. Default of 1
returns all partners. This threshold will be applied to the ROI weight when
the \code{roi} argument is specified, otherwise to the whole neuron.}

\item{roi}{a single ROI. Use \code{neuprint_ROIs} to see what is available.}

\item{by.roi}{logical, whether or not to break neurons' connectivity down by
region of interest (ROI)}

\item{conn}{Optional, a \code{neuprint_connection} object. Defaults to
\code{\link{fish_neuprint}} to ensure that the query targets fish2.}

\item{...}{Additional arguments passed to
\code{\link[neuprintr]{neuprint_connection_table}}.}

\item{dataset}{optional, a dataset you want to query. If \code{NULL}, the
default specified by your R environ file is used or, failing that the
current connection, is used. See \code{\link[neuprintr]{neuprint_login}} for details.}
}
\value{
A data.frame.
}
\description{
Connectivity query for fish2 neurons
}
\examples{
\donttest{
fish_connection_table("RGC", partners = "outputs")
fish_connection_table("RGC", partners = "outputs", summary = TRUE)
}
}
\seealso{
\code{\link[malevnc]{manc_connection_table}}

Other neuprint: 
\code{\link{fish_neuprint}()},
\code{\link{fish_neuprint_meta}()}
}
\concept{neuprint}
