% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprint.R
\name{fish_neuprint_meta}
\alias{fish_neuprint_meta}
\title{Fetch neuprint metadata for fish2 neurons}
\usage{
fish_neuprint_meta(
  ids = NULL,
  conn = NULL,
  roiInfo = FALSE,
  simplify.xyz = TRUE,
  ...,
  dataset = fish_default_dataset()
)
}
\arguments{
\item{ids}{Body ids, a query string (see \code{\link{fish_ids}}), or
\code{NULL} to return all bodies known to neuprint.}

\item{conn}{Optional, a \code{neuprint_connection} object. Defaults to
\code{\link{fish_neuprint}} to ensure that the query targets fish2.}

\item{roiInfo}{Whether to include ROI information (default \code{FALSE}).}

\item{simplify.xyz}{Whether to simplify columns containing XYZ locations to a
simple \code{"x,y,z"} format (default \code{TRUE}).}

\item{...}{Additional arguments passed to
\code{\link[neuprintr]{neuprint_get_meta}}.}

\item{dataset}{The name of the dataset as reported in Clio (default
\code{"fish2"}).}
}
\value{
A data.frame with one row for each unique input id and \code{NA}s for
  all columns except \code{bodyid} when neuprint holds no metadata.
}
\description{
Fetch neuprint metadata for fish2 neurons
}
\examples{
\donttest{
head(fish_neuprint_meta("RGC"))
}
}
\seealso{
\code{\link[malevnc]{manc_neuprint_meta}}

Other data-queries: 
\code{\link{fish_connection_table}()},
\code{\link{fish_ids}()},
\code{\link{fish_somapos}()},
\code{\link{register_fish_coconat}()}
}
\concept{data-queries}
