% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprint.R
\name{fish_roi_meshes}
\alias{fish_roi_meshes}
\title{Fetch one or more ROI meshes for fish2}
\usage{
fish_roi_meshes(
  rois,
  units = c("nm", "raw", "microns"),
  OmitFailures = TRUE,
  conn = NULL,
  ...,
  dataset = fish_default_dataset()
)
}
\arguments{
\item{rois}{Character vector of ROI names.}

\item{units}{One of \code{"nm"} (default), \code{"raw"} (8 nm voxels), or
\code{"microns"}.}

\item{OmitFailures}{Whether to omit ROI meshes that could not be read from
the server (default \code{TRUE}).}

\item{conn}{Optional, a \code{neuprint_connection} object. Defaults to
\code{\link{fish_neuprint}} to ensure that the query targets fish2.}

\item{...}{Additional arguments passed to \code{\link[nat]{nlapply}}.
This allows progress reporting and simple parallelisation.}

\item{dataset}{The name of the dataset as reported in Clio (default
\code{"fish2"}).}
}
\value{
A \code{shapelist3d} containing one or more ROI \code{mesh3d}
  objects.
}
\description{
Fetch one or more ROI meshes for fish2
}
\examples{
\dontrun{
rois <- fish_rois()
fish_roi_meshes(unique(c(rois$from, rois$to))[1:2])
}
}
\seealso{
\code{\link[neuprintr]{neuprint_ROI_mesh}},
  \code{\link[nat]{nlapply}},
  \code{\link{fish_rois}}

Other neuprint: 
\code{\link{fish_connection_table}()},
\code{\link{fish_neuprint}()},
\code{\link{fish_neuprint_meta}()},
\code{\link{fish_rois}()}
}
\concept{neuprint}
