% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprint.R
\name{fish_roi_meshes}
\alias{fish_roi_meshes}
\title{Fetch one or more ROI meshes for fish2}
\usage{
fish_roi_meshes(
  rois,
  units = c("nm", "raw", "microns"),
  conn = NULL,
  ...,
  dataset = fish_default_dataset()
)
}
\arguments{
\item{rois}{Character vector of ROI names.}

\item{units}{One of \code{"nm"} (default), \code{"raw"} (voxels), or
\code{"microns"}.}

\item{conn}{Optional, a \code{neuprint_connection} object. Defaults to
\code{\link{fish_neuprint}} to ensure that the query targets fish2.}

\item{...}{Additional arguments passed to \code{\link[nat]{nlapply}}. This
allows progress reporting and simple parallelisation.}

\item{dataset}{The name of the dataset as reported in Clio (default
\code{"fish2"}).}
}
\value{
A \code{shapelist3d} containing one or more ROI \code{mesh3d}
  objects. Missing meshes are dropped with a warning.
}
\description{
Fetch one or more ROI meshes for fish2
}
\details{
It seems that the nominal voxel dimensions of the dataset are 16 x
16 x 15 nm. ROI meshes are returned from neuprint in raw voxel space and are
scaled up by this factor to give the default nm scaling.
}
\examples{
\donttest{
fish_roi_meshes(c("Midbrain", "Hindbrain"), .progress = "none")
}
}
\seealso{
\code{\link[neuprintr]{neuprint_ROI_mesh}},
  \code{\link[nat]{nlapply}}, \code{\link{fish_rois}}

Other 3d-meshes-skeletons: 
\code{\link{fish_rois}()},
\code{\link{read_fish_meshes}()},
\code{\link{read_fish_neurons}()}
}
\concept{3d-meshes-skeletons}
