% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishr-package.R
\docType{package}
\name{fishr-package}
\alias{fishr}
\alias{fishr-package}
\title{fishr: Access the Fish2 FlyEM Dataset}
\description{
A thin wrapper around the malevnc package providing access to the fish2 larval zebrafish EM dataset being proofread at Janelia Research Campus as part of a collaboration with the Engert / Lichtman labs at Harvard and Google Research. Currently provides convenience functions for reading meshes and setting dataset options.
}
\section{Package Options}{


\itemize{

  \item \code{fishr.dataset} The name of the active fish2 dataset (default
  \code{"fish2"}). Set by \code{\link{choose_fish_dataset}}.

  }

  The following \code{malevnc.*} options are set by \code{\link{choose_fish}}
  or \code{\link{with_fish}} to point the \code{malevnc} package at the fish2
  servers:

  \itemize{

  \item \code{malevnc.server} DVID server URL (set via Clio lookup).

  \item \code{malevnc.rootnode} DVID root node UUID (set via Clio lookup).

  \item \code{malevnc.neuprint} neuprint server URL (fallback:
  \code{https://neuprint-fish2.janelia.org}).

  \item \code{malevnc.neuprint_dataset} neuprint dataset name (fallback:
  \code{"fish2"}).

  }
}

\examples{
\donttest{
options()[grepl("^fishr", names(options()))]
}
\dontrun{
with_fish(read_fish_meshes(12345))
}
}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

}
\keyword{internal}
