% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishr-package.R
\docType{package}
\name{fishr-package}
\alias{fishr}
\alias{fishr-package}
\title{fishr: Calcualte standard catch metrics for fisheries}
\description{
Provides functions for calculating and analyzing fisheries catch data, such as Catch Per Unit Effort (CPUE), a fundamental metric in fisheries science.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jacampbell0/fishr}
  \item Report bugs at \url{https://github.com/jacampbell0/fishr/issues}
}

}
\author{
\strong{Maintainer}: Jill Campbell \email{jill.campbell@dfo-mpo.gc.ca}

}
\keyword{internal}
