% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishr-package.R
\docType{package}
\name{fishr-package}
\alias{fishr}
\alias{fishr-package}
\title{fishr: Calculate standard catch metrics for fisheries}
\description{
Provides functions for analyzing fisheries data for fisheries scientists.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Jennifer-Boldt/fishr}
  \item Report bugs at \url{https://github.com/Jennifer-Boldt/fishr/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Boldt \email{Jennifer.Boldt@dfo-mpo.gc.ca}

}
\keyword{internal}
