% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishr-package.R
\docType{package}
\name{fishr-package}
\alias{fishr}
\alias{fishr-package}
\title{fishr: Calculate Standard Catch Metrics for Fisheries}
\description{
Provides functions for analyzing fisheries data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jjshao/fishr}
  \item Report bugs at \url{https://github.com/jjshao/fishr/issues}
}

}
\author{
\strong{Maintainer}: Jillian Shao \email{jillian.j.shao@gmail.com}

}
\keyword{internal}
