% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishr-package.R
\docType{package}
\name{fishr-package}
\alias{fishr}
\alias{fishr-package}
\title{fishr: Calculate standard catch metrics for fisheries}
\description{
Provides functions for calculating and analyzing fisheries catch data, such as Catch Per Unit Effort (CPUE), a fundamental metric in fisheries science.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/laurentjgill/fishr}
  \item Report bugs at \url{https://github.com/laurentjgill/fishr/issues}
}

}
\author{
\strong{Maintainer}: Lauren Gill \email{laurengill27@gmail.com} (\href{https://orcid.org/XXXX-XXXX-XXXX-XXXX}{ORCID})

}
\keyword{internal}
