% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishr-package.R
\docType{package}
\name{fishr-package}
\alias{fishr}
\alias{fishr-package}
\title{fishr: What the Package Does (One Line, Title Case)}
\description{
Functions to calculate typical fish and fisheries stats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lukewarkentin/fishr}
  \item Report bugs at \url{https://github.com/lukewarkentin/fishr/issues}
}

}
\author{
\strong{Maintainer}: Luke Warkentin \email{first.last@example.com}

}
\keyword{internal}
