% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishr-package.R
\docType{package}
\name{fishr-package}
\alias{fishr}
\alias{fishr-package}
\title{fishr: Calculate standard catch metrics for fisheries}
\description{
Who the heck knows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nickcfisch/fishr}
  \item Report bugs at \url{https://github.com/nickcfisch/fishr/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Fisch \email{nickcfisch@gmail.com}

}
\keyword{internal}
