% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishr-package.R
\docType{package}
\name{fishr-package}
\alias{fishr}
\alias{fishr-package}
\title{fishr: Calculate standard catch metrics for fisheries}
\description{
An example package developed during a TESA-led training on advanced R package development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ricardd/fishr}
  \item Report bugs at \url{https://github.com/ricardd/fishr/issues}
}

}
\author{
\strong{Maintainer}: Daniel Ricard \email{Daniel.Ricard@dfo-mpo.gc.ca}

}
\keyword{internal}
