% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{organize}
\alias{organize}
\title{}
\usage{
organize(fit, par.name, mcmc.out = FALSE)
}
\arguments{
\item{fit}{Model object from JAGS or Stan}

\item{par.name}{Parameter of intrest to organize into a dataframe}

\item{mcmc.out}{Logical, whether the full dataframe with all values should be
returned. Defaults to FALSE.}
}
\description{

}
\examples{
Add these in...
}
\author{
Michael J. Dodrill, \email{mdodrill@usgs.gov}
}
