% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{read_fish_meshes}
\alias{read_fish_meshes}
\title{Read meshes for fish2 body ids}
\usage{
read_fish_meshes(ids, node = "neutu", units = c("nm", "raw", "microns"), ...)
}
\arguments{
\item{ids}{One or more body ids (numeric or character), or a query string
accepted by \code{\link{fish_ids}}.}

\item{node}{The DVID node (UUID) to query. The default value of 'neutu' uses
the active neutu node i.e. normally the most up to date.}

\item{units}{One of \code{"nm"} (default), \code{"raw"} (8 nm voxels), or
\code{"microns"}.}

\item{...}{Additional arguments passed to \code{httr::GET}.}
}
\value{
A \code{\link[nat:neuronlist]{neuronlist}} containing one or more
  \code{mesh3d}
  objects.
}
\description{
Fetches pre-computed neuroglancer meshes from the fish2 DVID
  server. Meshes are returned in nm coordinates by default (8 nm voxels).
}
\examples{
\dontrun{
# pick 10 RGC meshes to read in
rgcs <- read_fish_meshes(fish_ids('RGC')[1:10])
plot3d(rgcs)
}
}
\seealso{
Other fishr-package: 
\code{\link{fish_dvid_annotations}()},
\code{\link{fish_ids}()},
\code{\link{fish_setup}()},
\code{\link{with_fish}()}
}
\concept{fishr-package}
