% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurons.R
\name{read_fish_neurons}
\alias{read_fish_neurons}
\title{Read neuron skeletons via neuprint}
\usage{
read_fish_neurons(
  ids,
  connectors = FALSE,
  units = c("nm", "raw", "microns"),
  heal.threshold = Inf,
  conn = NULL,
  ...,
  dataset = fish_default_dataset()
)
}
\arguments{
\item{ids}{Body ids in any form compatible with \code{\link{fish_ids}}.}

\item{connectors}{Whether to fetch synaptic connections for the neuron
(default \code{FALSE} in contrast to
\code{\link[neuprintr]{neuprint_read_neurons}}).}

\item{units}{Units of the returned neurons (default \code{"nm"}).}

\item{heal.threshold}{The threshold for healing disconnected skeleton
fragments. The default of \code{Inf} ensures that all fragments are joined
together.}

\item{conn}{Optional, a \code{neuprint_connection} object. Defaults to
\code{\link{fish_neuprint}} to ensure that the query targets fish2.}

\item{...}{Additional arguments passed to
\code{\link[neuprintr]{neuprint_read_neurons}}.}

\item{dataset}{The name of the dataset as reported in Clio (default
\code{"fish2"}).}
}
\value{
A \code{\link[nat:neuronlist]{neuronlist}} object containing one or
  more neurons.
}
\description{
Read neuron skeletons via neuprint
}
\examples{
\donttest{
rgcsk3 <- read_fish_neurons(fish_ids("RGC")[1:3])
plot(rgcsk3, WithNodes=FALSE)
}
}
\seealso{
\code{\link[malevnc]{manc_read_neurons}}, and nat functions
  including \code{\link[nat]{neuron}}, \code{\link[nat]{neuronlist}}.

Other fishr-package: 
\code{\link{fish_annotate}()},
\code{\link{fish_dvid_annotations}()},
\code{\link{fish_ids}()},
\code{\link{fish_setup}()},
\code{\link{read_fish_meshes}()},
\code{\link{with_fish}()}
}
\concept{fishr-package}
