% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{flyem_shorten_url}
\alias{flyem_expand_url}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{malevnc}{\code{\link[malevnc:flyem_shorten_url]{flyem_expand_url}}, \code{\link[malevnc]{flyem_shorten_url}}}
}}

