% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_trace.R
\name{stan_trace}
\alias{stan_trace}
\title{Trace plots for Stan model objects}
\usage{
stan_trace(fit, par.name, number, same.scale = FALSE)
}
\arguments{
\item{fit}{Stan model object.}

\item{par.name}{Name of the parameter of interest in quotes.}

\item{number}{A numeric vector giving the index for what values of the
parameter to display.  This can be also be a list for specifying paramaters
in a matrix. See Details.}

\item{same.scale}{Display the density plots on the same x-axis scale, across parameters. Defaults to false.}
}
\description{
Creates trace and density plots for Stan model objects, with
  R-hat and Neff displayed.
}
\examples{
Need to add examples.
}
\author{
Michael J. Dodrill, \email{mdodrill@usgs.gov}
}
